Moved UserInterface, created new entry point
Moved UserInterface class to the userinterface package. Started on some
changes for the language selection panel.
