+++ src/linecounter/userinterface/LanguageSelectionPanel.java
@@ -1,9 +1,12 @@
 package linecounter.userinterface;
 
+import java.awt.GridLayout;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
+import javax.swing.JButton;
 import javax.swing.JComboBox;
+import javax.swing.JTable;
 import javax.swing.border.TitledBorder;
 
 import linecounter.RunResult;
@@ -19,15 +22,22 @@ import linecounter.Settings;
 public class LanguageSelectionPanel extends SourceCounterPanel {
 	
 	private JComboBox<?> languagesDropDown;
+	private JTable detailsTable;
+	private JButton newLangButton;
 	
 	public LanguageSelectionPanel(Settings set) {
 		
 		this.settings = set;
 		
+		setLayout(new GridLayout(3, 1));
+		
 		languagesDropDown  = new JComboBox<Object>(settings.availableLangs.availableLangs());
         languagesDropDown.setSelectedIndex(settings.selectedLangIndex);
         
-     // LANGUAGES LIST DROPDOWN
+        detailsTable = new JTable();
+        
+        newLangButton = new JButton("Define New Language");
+        
         languagesDropDown.addActionListener(new ActionListener(){
             public void actionPerformed(ActionEvent ev){
                 // Update the settings
@@ -36,6 +46,8 @@ public class LanguageSelectionPanel extends SourceCounterPanel {
         });
         
         add(languagesDropDown);
+        add(detailsTable);
+        add(newLangButton);
         setBorder(new TitledBorder("Language"));
 	}
